refactor of directory
