+++ src/components/profile/NavigationSmartphone.tsx
